airflow-docker/requirements.txt
apache-airflow-providers-apache-spark==5.5.0
pyspark
faker
apache-airflow-providers-postgres